+++ src/main/java/openmods/geometry/Box2d.java
@@ -4,16 +4,16 @@ public class Box2d {
 
 	public static final Box2d NULL = new Box2d(0, 0, 0, 0, 0, 0);
 
-	public final int top;
-	public final int bottom;
+	public final float top;
+	public final float bottom;
 
-	public final int left;
-	public final int right;
+	public final float left;
+	public final float right;
 
-	public final int width;
-	public final int height;
+	public final float width;
+	public final float height;
 
-	private Box2d(int top, int bottom, int left, int right, int width, int height) {
+	private Box2d(float top, float bottom, float left, float right, float width, float height) {
 		this.top = top;
 		this.bottom = bottom;
 		this.left = left;
@@ -22,15 +22,15 @@ public class Box2d {
 		this.height = height;
 	}
 
-	public static Box2d fromCoords(int top, int bottom, int left, int right) {
+	public static Box2d fromCoords(float top, float bottom, float left, float right) {
 		if (bottom < top) {
-			final int tmp = bottom;
+			final float tmp = bottom;
 			bottom = top;
 			top = tmp;
 		}
 
 		if (left > right) {
-			final int tmp = left;
+			final float tmp = left;
 			left = right;
 			right = tmp;
 		}
@@ -38,9 +38,9 @@ public class Box2d {
 		return new Box2d(top, bottom, left, right, right - left, bottom - top);
 	}
 
-	public static Box2d fromOriginAndSize(int x, int y, int width, int height) {
-		final int left;
-		final int right;
+	public static Box2d fromOriginAndSize(float x, float y, float width, float height) {
+		final float left;
+		final float right;
 		if (width >= 0) {
 			left = x;
 			right = x + width;
@@ -50,8 +50,8 @@ public class Box2d {
 			width = -width;
 		}
 
-		final int top;
-		final int bottom;
+		final float top;
+		final float bottom;
 		if (height >= 0) {
 			bottom = y + height;
 			top = y;
@@ -64,7 +64,7 @@ public class Box2d {
 		return new Box2d(top, bottom, left, right, width, height);
 	}
 
-	public boolean isInside(int x, int y) {
+	public boolean isInside(float x, float y) {
 		final float dx = x - left;
 		final float dy = y - top;
